Refactor: Modularized project into src/ layout and integrated pytest-benchmark suite
